% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR Diagram.R
\name{import_mcdonald_data}
\alias{import_mcdonald_data}
\title{McDonald Observatory Dataset Import}
\usage{
import_mcdonald_data(filename)
}
\arguments{
\item{filename}{Filename in the Excel format to be read. The file must be located in the working directory.}
}
\value{
Returns a data frame from an Excel file into a dataframd called "mcdonaldstars".
}
\description{
This function imports an Excel file based on the McDonald Observatory data set.
}
\details{
Note that the McDonald Observatory data set is an updated data set that includes right ascension, declination, spectral types and general notes.
Follow the instructions on the NOAA site for downloading the entire dataset to a tab delimited file.
Star data from tables B and C in the McDonald Observatory data set were used as the basis of this data set.
\url{https://mcdonaldobservatory.org/sites/default/files/pdfs/teachers/hrd.pdf}
}
\examples{
# Downloaded mcdonaldstars.xlsx file to your working directory.
# Returns a dataframe with the default name "mcdonaldstar"
\dontrun{

library(readxl)
df<-import_mcdonald_data("mcdonaldstars.xlsx")

}
}
