% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR Diagram.R
\name{interactiveHR}
\alias{interactiveHR}
\title{Simple Interactive H-R Diagram (McDonald Observatory Dataset)}
\usage{
interactiveHR(filename)
}
\value{
An interactive plot with stars of differing temperatures and absolute magnitudes.  Star postions are given as well in the dataset.
}
\description{
This function creates an interactive (theoretical) H-R diagram on a small dataset (n=55 stars) from McDonald Observatory.
}
\note{
This graph is based on absolute magnitude and degrees Kelvin of the stars plotted.
}
\examples{
\dontrun{

# Use the dataframe name from the import routine.

library(dplyr)
library(ggplot2)
library(plotly)
interactiveHR(filename) #Filename will be the imported McDonald Observatory Dataset.

}   

}
