% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR Diagram.R
\name{spectral}
\alias{spectral}
\title{Convert stellar temperatures to spectral types based on degrees Kelvin.}
\usage{
spectral(filename)
}
\arguments{
\item{filename}{Filename is the data frame to be manipulated.}
}
\value{
Returns a data frame with spectral types.
}
\description{
This function adds spectral types to a dataframe given a column of stellar temperatures.
}
\note{
This function is generic and can be used to provide spectral types to any dataframe with a column of stellar temperatures in degrees Kelvin.
}
\examples{
\dontrun{
library(dplyr)
df<-spectral(filename) 

}  
}
