% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR Diagram.R
\name{yale_hr_diagram}
\alias{yale_hr_diagram}
\title{Hertzsprung-Russell Diagram (Yale Trigonometric Parallax Dataset)}
\usage{
yale_hr_diagram()
}
\value{
Returns a H-R plot based on absolute magnitude and color of stars.
}
\description{
This function creates an observational H-R diagram on a dataset from the Yale Trigonometric Parallax Dataset.
}
\note{
The data comes from the R package GDAdata which must be installed.
}
\examples{
\dontrun{

# Use the dataframe name from the import routine.

library(dplyr)
library(ggplot2)
library(plotly)
install.packages("GDAdata")
data(HRstars, package="GDAdata")
yale_hr_diagram

}   

}
