% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR Diagram.R
\name{interactiveHR}
\alias{interactiveHR}
\title{Simple Interactive H-R Diagram (McDonald Observatory Dataset)}
\usage{
interactiveHR(filename)
}
\value{
An interactive plot with stars of differing temperatures and absolute magnitudes.  Star postions are given as well in the dataset.
}
\description{
This function creates an interactive (theoretical) H-R diagram on a small dataset (n=55 stars) from McDonald Observatory.
}
\note{
The interactive plot also provides a tooltip that can be used to hover over individual stars to show additional information such as apparent magnitude, spectral type, constellation, right ascension and declination.
}
\examples{
\dontrun{

# Use the dataframe name from the import routine.

library(dplyr)
library(ggplot2)
library(plotly)
interactiveHR(filename) #Filename will be the imported McDonald Observatory Dataset.

}   

}
\references{
The exercises that accompanies this data may be found at https://mcdonaldobservatory.org/sites/default/files/pdfs/teachers/hrd.pdf
}
