% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR Diagram.R
\name{interactiveHR}
\alias{interactiveHR}
\title{Creating an Interactive H-R Diagram}
\usage{
interactiveHR(filename)
}
\value{
An interactive plot with stars of differing temperatures and absolute magnitudes.  Star postions are given as well.
}
\description{
This function creates an interactive H-R diagram on a limited dataset from the McDonald Observatory.
}
\note{
The routine also assigns spectral types to the McDonald data set.
}
\examples{
\dontrun{

# Use the dataframe name from the import routine.

library(dplyr)
library(ggplot2)
library(plotly)
interactiveHR(filename)

}   

}
